Fix server found/lost detection in stub middleware
The poll() call needs to monitor the inotify file-descriptor retuned by
the inotify_init() call instead of the watch-descriptor returned by the
inotify_add_watch() call. All inotify events came from the same
file-descriptor and we use the watch-descriptor to figure what watcher
has been notified.

Also keeps the PID file open by the server process until the process
dies or the server explicit calls UnpublishServer().

On the client side, in the first Poll() call  we try to read the server
PID file to discover if the server found event has already happened.

This commit has a lot of print messages, keep it here for a while, it'll
be useful in the development process.

Signed-off-by: Filipe Utzig <filipeutzig@gmail.com>