Fix MessageQueue move issues
In some instantiations of MessageQueus we are performing move operations
that calls the destructor of the original object, closing the mqueue in
the system, this is causing the poll() call to failed. Fixing this
problem with a proper move constructor/assignment operator
implementation that keeps de mqueue open and invalidate the old object.

Also add some useful assert macros.

Signed-off-by: Filipe Utzig <filipeutzig@gmail.com>